requirements/dev.txt
-r base.txt

# Testing dependencies
pytest>=7.0.0,<8.0.0
pytest-django>=4.5.2,<5.0.0
pytest-cov>=3.0.0,<4.0.0

# Linting and formatting tools
black>=22.1.0,<23.0.0
isort>=5.10.1,<6.0.0
flake8>=4.0.1,<5.0.0
pre-commit>=2.17.0,<3.0.0

# Debugging dependencies
django-debug-toolbar>=3.2.4,<4.0.0